+++ rust/src/datatypes/sequences/aamap/aaweights.rs
@@ -10,7 +10,7 @@ use rand::seq::WeightError;
 /// 1. Be non-negative, finite, and not NAN
 /// 2. Not be all zero
 /// 3. Have a finite sum
-pub struct AAWeights(AAMap<f32>);
+pub struct AAWeights(AAMap<f64>);
 impl AAWeights {
     /// Make a new [`AAWeights`].
     ///
@@ -18,7 +18,7 @@ impl AAWeights {
     /// 1. Any of the weights are negative or NAN ([`WeightError::InvalidWeight`]).
     /// 2. The sum of the weights reaches infinity ([`WeightError::Overflow`]).
     /// 3. All the weights are zero ([`WeightError::InsufficientNonZero`]).
-    pub fn new(weights: &AAMap<f32>) -> Result<&Self, WeightError> {
+    pub fn new(weights: &AAMap<f64>) -> Result<&Self, WeightError> {
         let mut total_weight = 0.0;
         for weight in weights.values().copied() {
             if !(weight >= 0.0) {
@@ -33,12 +33,12 @@ impl AAWeights {
             return Err(WeightError::InsufficientNonZero);
         }
         // SAFETY: [`AAWeights`] is just a newtype wrapper around [`AAMap`].
-        let this = unsafe { mem::transmute::<&AAMap<f32>, &Self>(weights) };
+        let this = unsafe { mem::transmute::<&AAMap<f64>, &Self>(weights) };
         Ok(this)
     }
 }
 impl Deref for AAWeights {
-    type Target = AAMap<f32>;
+    type Target = AAMap<f64>;
     fn deref(&self) -> &Self::Target {
         &self.0
     }
