+++ rust/src/datatypes/webworker_messages/common/featurize.rs
@@ -5,22 +5,22 @@ use crate::datatypes::StandardError;
 #[derive(Deserialize, Serialize)]
 #[serde(tag = "case")]
 pub enum Featurized {
-    Ok { value: f32 },
+    Ok { value: f64 },
     /// Error coming from featurization
     /// of a single sequence.
     Error { value: StandardError },
 }
 impl Featurized {
     /// Get the feature value if it was successfully computed.
-    pub fn ok(&self) -> Option<f32> {
+    pub fn ok(&self) -> Option<f64> {
         match *self {
             Self::Ok { value } => Some(value),
             Self::Error { .. } => None
         }
     }
 }
-impl From<Result<f32, StandardError>> for Featurized {
-    fn from(value: Result<f32, StandardError>) -> Self {
+impl From<Result<f64, StandardError>> for Featurized {
+    fn from(value: Result<f64, StandardError>) -> Self {
         match value {
             Ok(value) => Self::Ok {value},
             Err(err) => Self::Error { value: err }
