+++ rust/src/seq_features/implementors/log_ratio/featurize.rs
@@ -16,7 +16,7 @@ impl FeaturizableSeqFeats for LogRatioContainer {
         &self,
         _sequence: &crate::datatypes::aa_canonical_str,
         ctx: Self::Ctx<'a>,
-    ) -> impl Iterator<Item = Result<f32, Self::Err>> {
+    ) -> impl Iterator<Item = Result<f64, Self::Err>> {
         self.data
             .iter()
             .map(|feature| Ok(compute_soft_log_ratio(feature, ctx)))
@@ -25,10 +25,10 @@ impl FeaturizableSeqFeats for LogRatioContainer {
 /// Helper function for [`LogRatioContainer::featurize`].
 ///
 /// Compute log1p ratio (natural log).
-fn compute_soft_log_ratio(feature: &LogRatio, residue_counts: &ResidueCounts) -> f32 {
+fn compute_soft_log_ratio(feature: &LogRatio, residue_counts: &ResidueCounts) -> f64 {
     let LogRatio {
         numerator,
         denominator,
     } = *feature;
-    ((residue_counts[numerator] + 1) as f32 / (residue_counts[denominator] + 1) as f32).ln()
+    ((residue_counts[numerator] + 1) as f64 / (residue_counts[denominator] + 1) as f64).ln()
 }
